Fix id type in db

+++ thentos-core/schema/schema.sql
@@ -8,6 +8,6 @@ CREATE TABLE IF NOT EXISTS "users" (
 
 CREATE TABLE IF NOT EXISTS "password_reset_tokens" (
     token      text      NOT NULL,
-    uid        integer   REFERENCES users (id),
+    uid        bigint    REFERENCES users (id),
     timestamp  timestamp DEFAULT now()
 );
